refactor |mailmsghdrctype|
partial refactor; more needed
